Re-organise utility and unnused functions

+++ includes/deploy_growing_guide.php
@@ -31,6 +31,35 @@ define('SKIP_GROWING_RESOURCE_IDS', [
 
 
 if (defined('WP_CLI') && WP_CLI) {
+
+    /**
+     * Registers a WP-CLI command `vs set_category_growersguide` to associate a growing guide with a product category.
+     *
+     * Command Usage:
+     *   wp vs set_category_growersguide <category_id> <guide_id>
+     *
+     * Parameters:
+     *   @param int $cat_id   The ID of the product category to associate with the growing guide.
+     *   @param int $guide_id The ID of the growing guide (post) to associate with the category.
+     */
+    WP_CLI::add_command('vs set_category_growersguide', function($cat_id, $guide_id ) {
+        $guide = get_post($guide_id);
+        if ($guide) {
+            WP_CLI::line("Guide Title: " . $guide->post_title . "\n");
+        } else {
+            WP_CLI::line("No page found for guide ID: $guide_id\n");
+        }
+
+        $category = get_term($cat_id, 'product_cat');
+        if ($category && !is_wp_error($category)) {
+            WP_CLI::line("Category Name: " . $category->name . "\n");
+        } else {
+            WP_CLI::line("No category found for category ID: $cat_id\n");
+        }
+        WP_CLI::confirm('Update the growing guide field for this category?', ['y', 'n']);
+        update_field('growing_guide', $guide_id, 'term_' . $cat_id);
+        // update_field('growing_guide', $page->ID, 'term_' . $term->term_id);
+    });
     WP_CLI::add_command('vs list_growing_guides', function() {
         $guides = get_posts([
             'post_type' => 'growing-guide',
@@ -196,7 +225,6 @@ function remove_empty_parent_categories($tree, $parent_id = 0, $depth = 0) {
     }
 }
 
-
 function delete_all_growing_guides() {
     $posts = get_posts([
         'post_type' => 'growing-guide',
@@ -235,151 +263,6 @@ function create_empty_growers_guides($terms) {
             echo "{$post_url}\n";
         }
     }
-}
-
-function description_headings_by_phrase($description, $phrase) {
-    $headings = get_headings_from_description($description);
-    $matching_headings = [];
-    foreach ($headings as $heading) {
-        if (is_array($phrase)) {
-            foreach ($phrase as $p) {
-                if (stripos($heading, $p) !== false) {
-                    $matching_headings[] = $heading;
-                    break;
-                }
-            }
-        } else {
-            if (stripos($heading, $phrase) !== false) {
-                $matching_headings[] = $heading;
-            }
-        }
-    }
-    return $matching_headings;
-}
-
-function description_has_growing_information($description) {
-    preg_match_all('/<h[1-6][^>]*>(.*?Growing.*?)<\/h[1-6]>/', $description, $matches);
-    return $matches[1];
-    // return description_headings_by_phrase($description, 'Growing');
-}
-
-function get_headings_from_description($description) {
-    // preg_match_all('/<h[1-6][^>]*>(.*?)<\/h[1-6]>/', $description, $matches);
-    // return $matches[1];
-    if (empty($description)) {
-        return [];
-    }
-    $dom = new DOMDocument();
-    libxml_use_internal_errors(true);
-    $dom->loadHTML($description);
-    libxml_clear_errors();
-    $xpath = new DOMXPath($dom);
-
-    $headings = [];
-    foreach (range(1, 6) as $level) {
-        $nodes = $xpath->query("//h{$level}");
-        foreach ($nodes as $node) {
-            $headings[] = trim($node->textContent);
-        }
-    }
-    return $headings;
-}
-
-function get_sections_between_headings($page_content, $headings) {
-    if (empty($page_content)) {
-        return [];
-    }
-    $sections = [];
-    $dom = new DOMDocument();
-    libxml_use_internal_errors(true);
-    $dom->loadHTML(mb_convert_encoding($page_content, 'HTML-ENTITIES', 'UTF-8'));
-    libxml_clear_errors();
-    $xpath = new DOMXPath($dom);
-
-    foreach ($headings as $key => $heading) {
-        if (!empty($heading)) {
-            foreach ($heading as $h) {
-                $section = '';
-                // Find nodes that follow the heading
-                $nodes = $xpath->query("//*[contains(text(), '$h')]/following-sibling::*");
-                foreach ($nodes as $node) {
-                    // Stop if another heading is encountered
-                    if (in_array(strtolower($node->nodeName), ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'])) {
-                        break;
-                    }
-                    // Append the node's HTML to the section
-                    $section .= $dom->saveHTML($node);
-                }
-                // Add the section to the sections array
-                $sections[$key] = $section;
-            }
-        }
-    }
-    return $sections;
-}
-
-function best_match_page_title($pages, $string, $extended_string=null) {
-    $best_match = null;
-    $best_similarity = 0;
-    $threshold = 40;
-    $extended_string = $extended_string ?? $string;
-    foreach ($pages as $page) {
-        similar_text($string, $page->post_title, $similarity);
-        // if the exact string is found in the title give it a boost
-        if (stripos($page->post_title, $extended_string) !== false) {
-            $similarity += 20;
-        }
-        if ($similarity > $best_similarity) {
-            $best_similarity = $similarity;
-            $best_match = $page;
-        }
-    }
-    if ($best_similarity < $threshold) {
-        $best_match = null;
-    }
-    return $best_match;
-}
-
-function best_match_category_title($terms, $string, $extended_string=null, $threshold=50) {
-    $best_match = null;
-    $best_similarity = 0;
-    $extended_string = $extended_string ?? $string;
-    foreach ($terms as $term) {
-        similar_text($string, $term->name, $similarity);
-        // if the exact string is found in the title give it a boost
-        if (stripos($term->post_title, $extended_string) !== false) {
-            $similarity += 20;
-        }
-        if ($similarity > $best_similarity) {
-            $best_similarity = $similarity;
-            $best_match = $term;
-        }
-    }
-    if ($best_similarity < $threshold) {
-        $best_match = null;
-    }
-    return $best_match;
-}
-
-function strip_extra_markup($content) {
-    $content = strip_tags($content, ['p', 'a', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'ul', 'ol', 'li']);
-    // remove class attributes
-    $content = preg_replace('/class="[^"]*"/', '', $content);
-    // remove target attributes
-    $content = preg_replace('/target="[^"]*"/', '', $content);
-    // Remove extra whitespace
-    $content = preg_replace('/\s+/', ' ', $content);
-    return $content;
-}
-
-function get_rendered_html($post) {
-    $post_id = is_object($post) ? $post->ID : $post;
-    $post = is_int($post) ? get_post($post) : $post;
-
-    // remove_filter('the_content', 'do_shortcode', 11);
-    $page = get_post($post_id);
-    $content = apply_filters('the_content', $post->post_content);
-    return $content;
 }
 
 function create_category_growers_guide_from_page($term, $page, $title=null) {
@@ -425,7 +308,6 @@ function create_category_growers_guide_from_page($term, $page, $title=null) {
         echo $section . "\n";;
         echo "----------------------------------------\n";
     }
-    return;
     $post_id = wp_insert_post([
         'post_title' => $title ?? 'How to grow ' . $term->name,
         'post_content' => $page->post_content,
@@ -447,10 +329,14 @@ function create_category_growers_guide_from_page($term, $page, $title=null) {
         update_field('images', $media_ids, $post_id);
         update_field('original_growing_reference_page', $page->ID, $post_id);
 
-
+        // Add the category to the growing guide
         wp_set_object_terms($post_id, $term->term_id, 'product_cat');
         echo "Created guide: \033[36m{$term->name} \033[0m -> \033[35m{$page->post_title} (guide)\033[0m\n";
         $growers_guide = get_post($post_id);
+
+        // Link this growing guide to the category for display
+        update_field('growing_guide', $page->ID, 'term_' . $term->term_id);
+
         get_permalink($growers_guide->ID);
         return $growers_guide;
     } else {
@@ -459,47 +345,10 @@ function create_category_growers_guide_from_page($term, $page, $title=null) {
     }
 }
 
-/**
- * Adds a redirection using the Redirection plugin's REST API.
- *
- * @param string $source_url The URL to redirect from.
- * @param string $target_url The URL to redirect to.
- * @param int $redirect_type The type of redirection (default is 301).
- * @return bool True if the redirection was added successfully, false otherwise.
- */
-function add_redirection($source_url, $target_url, $redirect_type = 301, $cli = true) {
-    $api_url = rest_url('redirection/v1/redirect');
-    $args = array(
-        'method' => 'POST',
-        'headers' => array(
-            'Content-Type' => 'application/json',
-            'X-WP-Nonce' => wp_create_nonce('wp_rest')
-        ),
-        'body' => json_encode(array(
-            'url' => $source_url,
-            'action_data' => array('url' => $target_url),
-            'action_type' => 'url',
-            'action_code' => $redirect_type,
-            'group_id' => 1, // Default group
-            'enabled' => true
-        ))
-    );
-    $response = wp_remote_post($api_url, $args);
-    if (is_wp_error($response)) {
-        return false;
+function create_growers_guide_from_resource_page($growing_resources_page_id, $check_only=false, $delete=True) {
+    if ($delete) {
+        delete_all_growing_guides();
     }
-    $body = wp_remote_retrieve_body($response);
-    $data = json_decode($body, true);
-    if (isset($data['id'])) {
-        "Added redirection from \033[33m$source_url\033[0m\n to \033[33m$target_url\033[0m\n";
-    } else {
-        echo "\033[31mFailed to add redirection from \033[33m$source_url\033[31m\n to \033[33m$target_url\033[31m\n\033[0m";
-    }
-    return isset($data['id']);
-}
-
-function create_growers_guide_from_resource_page($growing_resources_page_id, $check_only=false) {
-    delete_all_growing_guides();
     $resource_pages = get_pages(['child_of' => $growing_resources_page_id]);
     $terms = get_terms(['taxonomy' => 'product_cat', 'hide_empty' => false]);
     $categories = array_filter($terms, function($term) {
@@ -510,6 +359,9 @@ function create_growers_guide_from_resource_page($growing_resources_page_id, $ch
 
     foreach ($resource_pages as $page) {
         if ( in_array($page->ID, SKIP_GROWING_RESOURCE_IDS)) {
+            continue;
+        }
+        if ($page->post_title != 'How to grow basil') {
             continue;
         }
         $title = str_replace('How to grow ', '', $page->post_title);
@@ -536,133 +388,4 @@ function create_growers_guide_from_resource_page($growing_resources_page_id, $ch
             WP_CLI::log("\033[31m- {$page->post_title} ({$page->ID})\033[0m\n");
         }
     }
-}
-
-function create_growers_guides_from_product_cat($terms, $check_only = true) {
-    // ✔ step through by category
-    // ✔ look for matching growing guide resource
-    //    - if we have one
-    //        ✔ convert to a growers guide
-    //        ✔ add link to the category
-    //    - if we don't
-    //         - check the category description for growing information
-    //         - divide up the description by heading, use regex or simple explode
-    //         - create growers guide using headings as fields
-    //         - add link to the category
-    // - step through products
-    //      - check if product description matches category growing guide
-    //          - if not create a guide from product
-    //              - check the product descriptions
-    //              - compare product descriptions within category
-    //                  - if the same (or similar?) make category guide
-    //                  - else make product guides
-    // - log guides created
-    // - note categories and products that should be skipped
-    //      - eg don't create product guide even if not match category guide)
-
-    $count_seeds_categories = 0;
-    $count_seeds_category_growing_instructions = 0;
-
-    // delete_all_growing_guides();
-    echo "========================================\n\n";
-
-    foreach ($terms as $term) {
-
-        if (is_under_seeds_category($term->term_id) && $term->description) {
-            $cat_name = trim(str_replace(['Seeds', 'seeds'], '', $term->name));
-            $guide_title = 'How to grow ' . $cat_name;
-
-            // Don't create guides for categories with no products
-            $products_in_category = get_posts([
-                'post_type' => 'product',
-                'numberposts' => -1,
-                'post_status' => 'publish',
-                'tax_query' => [
-                    [
-                        'taxonomy' => 'product_cat',
-                        'field' => 'term_id',
-                        'terms' => $term->term_id,
-                        'include_children' => true,
-                    ],
-                ],
-            ]);
-            if (empty($products_in_category)) {
-                echo "No products found for category: \033[33m{$cat_name}\033[0m\n";
-                continue;
-            }
-
-            if ($cat_name !== 'Pea') {
-                continue;
-            }
-
-            $term_url = get_term_link($term);
-            $count_seeds_categories ++;
-
-            $growing_resource_pages = get_posts([
-                'post_type' => 'page',
-                's' => $guide_title,
-                'post_status' => 'publish',
-                'posts_per_page' => -1,
-            ]);
-
-            // Growing resources page
-            if (!empty($growing_resource_pages)) {
-                echo "Category: \033[33m{$cat_name}:\033[0m\n";
-                echo "----------------------------------------\n";
-
-                $best_match = best_match_page_title($growing_resource_pages, $cat_name, $guide_title);
-
-                echo "Existing guide page(s):\n";
-                foreach ($growing_resource_pages as $page) {
-                    $asterisk = ($page == $best_match) ? " *" : "  ";
-                    // resource page title contains category name
-                    if (stripos($page->post_title, $cat_name) !== false) {
-                        // $edit_link = get_edit_post_link($page->ID);
-                        $page_url = get_permalink($page->ID);
-                        echo "  - {$page->post_title} ({$page->ID}) $asterisk  - {$page_url}\n";
-                    }
-                }
-                echo "----------------------------------------\n";
-                if ($best_match) {
-                    $guide = create_category_growers_guide_from_page($term, $best_match, $guide_title);
-                    if ($guide) {
-                        update_field('growers_guide', $guide->ID, 'term_' . $term->term_id);
-                        $permalink = get_permalink($guide);
-                        echo $permalink . "\n";
-                        add_redirection(get_permalink($best_match), $permalink, 302);
-                    }
-                }
-                echo "========================================\n";
-            }
-
-            // Check the category description for growing information too
-            // if ($headings_including_growing = description_has_growing_information($term->description)) {
-            //     $count_seeds_category_growing_instructions ++;
-
-            //     if ($items_with_growing_info) {
-            //         $all_headings = get_headings_from_description($term->description);
-
-            //         echo $cat_name . "\n";
-            //         echo $term_url . "\n";
-            //         echo "--------------------------------------\n";
-            //         // foreach ($headings_including_growing as $heading) {
-            //         //     echo "  " . strip_tags($heading) . "\n";
-            //         // }
-            //         foreach ($all_headings as $heading) {
-            //             echo "  " . strip_tags($heading) . "\n";
-            //         }
-            //         // $description = wordwrap($term->description, 120, "\n    ");
-            //         // echo $description . "\n";
-            //         echo "======================================-\n";
-            //     }
-            // } else if (!$items_with_growing_info) {
-            //     echo $cat_name . "\n";
-            //     echo $term_url . "\n";
-            //     echo "--------------------------------------\n";
-
-            // }
-        }
-    }
-    echo "\n\nSeeds categories: {$count_seeds_categories}\n";
-    echo "Seeds category growing instructions: {$count_seeds_category_growing_instructions}\n";
 }
